refactor: Dashboard components 폴더 정리
